requirements.txt
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
idna==3.4
isodate==0.6.1
lxml==4.9.2
meemoo-cloudevents==0.1.0rc3
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pulsar-client==3.0.0
pyparsing==3.0.9
python-json-logger==2.0.7
PyYAML==6.0
rdflib==6.2.0
requests==2.28.2
six==1.16.0
structlog==22.3.0
tomli==2.0.1
urllib3==1.26.14
viaa-chassis==0.2.0rc1
